feat: Add Fruits page with images
Implemented the "Fruits" page with an HTML structure showcasing three paragraphs of fruit images. Used the <img> tag to display banana, orange, kiwi, and apple images. The page is designed to be responsive with appropriate meta tags in the head section. This commit marks the completion of the "Fruits" page in the HTML and CSS learning journey.