feat: Add content about The Brown Bear
Added content to "The Brown Bear" page, including information about the brown bear species, its conservation status, subspecies, and notable populations in various countries. Included a link to learn more about brown bears and displayed an image of a bear. Utilized HTML tags such as <p>, <a>, <ul>, <li>, <ol>, and <img> to structure and present the information. This commit represents the completion of the content for "The Brown Bear" page in the HTML and CSS learning journey.